+++ package.json
@@ -24,7 +24,6 @@
     "class-variance-authority": "^0.4.0",
     "clsx": "^1.2.1",
     "framer-motion": "^8.0.2",
-    "jotai": "^2.0.2",
     "next": "13.1.0",
     "next-auth": "^4.18.7",
     "react": "18.2.0",
