+++ src/components/Settings/Settings.tsx
@@ -3,7 +3,7 @@ import type { FC } from 'react';
 import { useEffect } from 'react';
 import { Controller, useForm, useWatch } from 'react-hook-form';
 import type { Settings as SettingsFormValues } from 'src/api/models/settings/settings';
-import { useSettings } from 'src/store/settingsAtom';
+import { useSettings } from 'src/hooks/useSettings';
 import { Checkbox } from '../Checkbox/Checkbox';
 import { Field } from '../Field/Field';
 import { Form } from '../Form/Form';
@@ -14,6 +14,7 @@ import { RadioGroup } from '../RadioGroup';
 /** Settings popover component. */
 export const Settings: FC = () => {
   const [settings, setSettings] = useSettings();
+
   const methods = useForm({
     defaultValues: settings,
   });
