+++ src/features/search/components/CharacterSearch/CharacterSearch.tsx
@@ -8,8 +8,8 @@ import { Pagination } from 'src/api/models/pagination';
 import { useForm, useWatch } from 'react-hook-form';
 import { Form } from 'src/components/Form/Form';
 import { useDebounce } from 'usehooks-ts';
-import { useSettings } from 'src/store/settingsAtom';
 import { CensorService } from 'src/api/services/censorService';
+import { useSettings } from 'src/hooks/useSettings';
 import { DEFAULT_PAGE_SIZE, useCharactersQuery } from '../../queries/characters';
 import { CharacterSearchFormValues, CHARACTER_SEARCH_INITIAL_VALUES } from '../CharacterSearchForm/characterSearchFormValues';
 import { CharacterSearchForm } from '../CharacterSearchForm/CharacterSearchForm';
