+++ src/features/search/components/CharacterSearchForm/CharacterSearchForm.tsx
@@ -15,8 +15,8 @@ import { IconButton } from 'src/components/IconButton/IconButton';
 import { Select } from 'src/components/Select';
 import { TextInput } from 'src/components/TextInput/TextInput';
 import { useDebounce } from 'usehooks-ts';
-import { useSettings } from 'src/store/settingsAtom';
 import { CensorService } from 'src/api/services/censorService';
+import { useSettings } from 'src/hooks/useSettings';
 import { useTraitsQuery } from '../../queries/traits';
 import { useVnsQuery } from '../../queries/vns';
 import { SearchPopover } from '../SearchPopover/SearchPopover';
