+++ src/features/search/components/VnSearch/VnSearch.tsx
@@ -8,8 +8,8 @@ import { CardSkeleton } from 'src/components/Card/CardSkeleton';
 import { Form } from 'src/components/Form/Form';
 import { ChildrenMultiplier } from 'src/components/ChildrenMultiplier/ChildrenMultiplier';
 import { Pagination } from 'src/api/models/pagination';
-import { useSettings } from 'src/store/settingsAtom';
 import { CensorService } from 'src/api/services/censorService';
+import { useSettings } from 'src/hooks/useSettings';
 import { DEFAULT_PAGE_SIZE, useVnsQuery } from '../../queries/vns';
 import { VnSearchFormValues, VN_SEARCH_INITIAL_VALUES } from '../VnSearchForm/vnSearchFormValues';
 import { VnSearchForm } from '../VnSearchForm/VnSearchForm';
