+++ src/features/search/components/VnSearchForm/VnSearchForm.tsx
@@ -18,8 +18,8 @@ import { Select } from 'src/components/Select';
 import { Slider } from 'src/components/Slider/Slider';
 import { TextInput } from 'src/components/TextInput/TextInput';
 import { useDebounce } from 'usehooks-ts';
-import { useSettings } from 'src/store/settingsAtom';
 import { CensorService } from 'src/api/services/censorService';
+import { useSettings } from 'src/hooks/useSettings';
 import { useTagsQuery } from '../../queries/tags';
 import { SearchPopover } from '../SearchPopover/SearchPopover';
 import type { VnSearchFormValues } from './vnSearchFormValues';
